feat: Add support for more file types
This commit extends the script to recognize and list more file types in the README.
Now the script will also list Javascript, Typescript, HTML, CSS and JSON files in addition to Python files.
This change allows the script to be more useful in projects that use a wider variety of file types.